# Speed up function `val2tuple` by 148%